rework test
