+++ e2e/bug-11857/.gitignore
@@ -1,0 +1,1 @@
+/vendor
