+++ e2e/bug-11857/composer.json
@@ -1,0 +1,5 @@
+{
+	"autoload-dev": {
+		"classmap": ["src/"]
+	}
+}
