+++ e2e/bug-11857/phpstan.neon
@@ -1,3 +1,8 @@
+parameters:
+	level: 8
+	paths:
+		- src
+
 services:
 	-
 		class: Bug11857\RelationDynamicMethodReturnTypeExtension
